fix: Moves code on app.js to specific components
